requirements.txt
Django
djangorestframework
drf-yqsg
djoser